backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langgraph
langchain-google-genai
pydantic[email]
python-dotenv
httpx
